Add files via upload
The four files contain Python 3 code for reading an image of a supermarket receipt and writing information out of it automatically to an Excel file. Only a few parameters have to be set, for example the path and filename of the picture. Also Tesseract has to be downloaded and installed, which allows Python to use the pytesseract library. The image of the receipt will be read and transformed into a text object. From this text object the purchased items, amounts, and prices will be extracted and stored into a Pandas dataframe. This dataframe will then be written into an Excel file. New information will be appended to the existing Excel file. The column width within the Excel sheet will be set automatically depending on the content length. The file "one_file_version" contains the whole code in one, the other files are module based with one main file, "Receipt_Price_Extractor", and two files where the functions used are defined. The development was done in Jupyter Notebooks.